# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aussie Horse Win API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrackState = typeof TrackState[keyof typeof TrackState];


export const TrackState = {
  VIC: 'VIC',
  NSW: 'NSW',
} as const;

export type TrackType = typeof TrackType[keyof typeof TrackType];


export const TrackType = {
  Regional: 'Regional',
  Provincial: 'Provincial',
} as const;

export interface Track {
  id: number;
  name: string;
  state: TrackState;
  type: TrackType;
  enabled: boolean;
}

export type RunnerSpeedMapPosition = typeof RunnerSpeedMapPosition[keyof typeof RunnerSpeedMapPosition];


export const RunnerSpeedMapPosition = {
  Lead: 'Lead',
  'On-Pace': 'On-Pace',
  Handy: 'Handy',
  Midfield: 'Midfield',
  'Back-Marker': 'Back-Marker',
} as const;

export interface FilterResult {
  rule: string;
  passed: boolean;
  message: string;
}

export interface Runner {
  id: number;
  raceId: number;
  horseName: string;
  barrierNumber: number;
  speedMapPosition: RunnerSpeedMapPosition;
  winOdds: number;
  placeOdds: number;
  /** @nullable */
  jockey?: string | null;
  /** @nullable */
  trainer?: string | null;
  passed: boolean;
  filterResults: FilterResult[];
}

export interface Race {
  id: number;
  trackName: string;
  state: string;
  raceNumber: number;
  /** @nullable */
  raceName?: string | null;
  raceDate: string;
  /** @nullable */
  raceTime?: string | null;
  fieldSize: number;
  /** @nullable */
  distance?: number | null;
  runners: Runner[];
  qualifiedCount: number;
}

export type NominationStatus = typeof NominationStatus[keyof typeof NominationStatus];


export const NominationStatus = {
  Pending: 'Pending',
  Won: 'Won',
  Placed: 'Placed',
  Unplaced: 'Unplaced',
} as const;

export interface Nomination {
  id: number;
  raceId: number;
  runnerId: number;
  trackName: string;
  state: string;
  raceNumber: number;
  /** @nullable */
  raceName?: string | null;
  raceDate: string;
  /** @nullable */
  raceTime?: string | null;
  horseName: string;
  barrierNumber: number;
  speedMapPosition: string;
  winOdds: number;
  placeOdds: number;
  winStake: number;
  placeStake: number;
  totalOutlay: number;
  projectedWinReturn: number;
  projectedPlaceReturn: number;
  /** @nullable */
  jockey?: string | null;
  /** @nullable */
  trainer?: string | null;
  status: NominationStatus;
}

export type NominationSummaryByTrackItem = {
  trackName: string;
  count: number;
};

export type NominationSummaryByStateItem = {
  state: string;
  count: number;
};

export interface NominationSummary {
  totalNominations: number;
  totalOutlay: number;
  byTrack: NominationSummaryByTrackItem[];
  byState: NominationSummaryByStateItem[];
  pendingCount: number;
  wonCount: number;
  placedCount: number;
  unplacedCount: number;
}

export interface Settings {
  fieldSizeMin: number;
  fieldSizeMax: number;
  minWinOdds: number;
  maxWinOdds: number;
  minPlaceOdds: number;
  winStake: number;
  placeStake: number;
  enabledTrackIds: number[];
}

export interface SettingsInput {
  fieldSizeMin?: number;
  fieldSizeMax?: number;
  minWinOdds?: number;
  maxWinOdds?: number;
  minPlaceOdds?: number;
  winStake?: number;
  placeStake?: number;
  enabledTrackIds?: number[];
}

export interface ResultInput {
  runnerId: number;
  finishPosition: number;
  /** @nullable */
  actualWinReturn?: number | null;
  /** @nullable */
  actualPlaceReturn?: number | null;
}

export type BetResultOutcome = typeof BetResultOutcome[keyof typeof BetResultOutcome];


export const BetResultOutcome = {
  Won: 'Won',
  Placed: 'Placed',
  Unplaced: 'Unplaced',
} as const;

export interface BetResult {
  id: number;
  nominationId: number;
  raceId: number;
  runnerId: number;
  trackName: string;
  raceDate: string;
  horseName: string;
  finishPosition: number;
  fieldSize: number;
  winStake: number;
  placeStake: number;
  totalOutlay: number;
  /** @nullable */
  actualWinReturn?: number | null;
  /** @nullable */
  actualPlaceReturn?: number | null;
  netResult: number;
  outcome: BetResultOutcome;
}

export interface Performance {
  totalBets: number;
  totalOutlay: number;
  totalReturns: number;
  netProfitLoss: number;
  winStrikeRate: number;
  placeStrikeRate: number;
  roi: number;
  avgOddsWin: number;
  avgOddsPlace: number;
  longestWinStreak: number;
  longestLosingStreak: number;
  totalWins: number;
  totalPlaced: number;
  totalUnplaced: number;
}

export interface TrackPerformance {
  trackName: string;
  state: string;
  totalBets: number;
  totalOutlay: number;
  totalReturns: number;
  netProfitLoss: number;
  winStrikeRate: number;
  placeStrikeRate: number;
  roi: number;
}

export interface SyncResult {
  racesFound: number;
  racesAdded: number;
  runnersAdded: number;
  nominationsGenerated: number;
  message: string;
}