# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aussie Horse Win API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /races/{id}:
    get:
      operationId: getRace
      tags: [races]
      summary: Get a single race with full runner details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race with runner filter results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
        "404":
          description: Race not found
  /races/{id}/result:
    post:
      operationId: recordResult
      tags: [races]
      summary: Record race result for a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultInput"
      responses:
        "200":
          description: Result recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResult"
...
components:
  schemas:
...
    Runner:
      type: object
      properties:
        id:
          type: integer
        raceId:
          type: integer
        horseName:
          type: string
        barrierNumber:
          type: integer
        speedMapPosition:
          type: string
          enum: [Lead, On-Pace, Handy, Midfield, Back-Marker]
        winOdds:
          type: number
        placeOdds:
          type: number
        jockey:
          type: ["string", "null"]
        trainer:
          type: ["string", "null"]
        passed:
          type: boolean
        filterResults:
          type: array
          items:
            $ref: "#/components/schemas/FilterResult"
      required:
        - id
        - raceId
        - horseName
        - barrierNumber
        - speedMapPosition
        - winOdds
        - placeOdds
        - passed
        - filterResults
    FilterResult:
      type: object
      properties:
        rule:
          type: string
        passed:
          type: boolean
        message:
          type: string
      required:
        - rule
        - passed
        - message
    Race:
      type: object
      properties:
        id:
          type: integer
        trackName:
          type: string
        state:
          type: string
        raceNumber:
          type: integer
        raceName:
          type: ["string", "null"]
        raceDate:
          type: string
        raceTime:
          type: ["string", "null"]
        fieldSize:
          type: integer
        distance:
          type: ["integer", "null"]
        runners:
          type: array
          items:
            $ref: "#/components/schemas/Runner"
        qualifiedCount:
          type: integer
      required:
        - id
        - trackName
        - state
        - raceNumber
        - raceDate
        - fieldSize
        - runners
        - qualifiedCount
...
    ResultInput:
      type: object
      properties:
        runnerId:
          type: integer
        finishPosition:
          type: integer
        actualWinReturn:
          type: ["number", "null"]
        actualPlaceReturn:
          type: ["number", "null"]
      required:
        - runnerId
        - finishPosition
    BetResult:
      type: object
      properties:
        id:
          type: integer
        nominationId:
          type: integer
        raceId:
          type: integer
        runnerId:
          type: integer
        trackName:
          type: string
        raceDate:
          type: string
        horseName:
          type: string
        finishPosition:
          type: integer
        fieldSize:
          type: integer
        winStake:
          type: number
        placeStake:
          type: number
        totalOutlay:
          type: number
        actualWinReturn:
          type: ["number", "null"]
        actualPlaceReturn:
          type: ["number", "null"]
        netResult:
          type: number
        outcome:
          type: string
          enum: [Won, Placed, Unplaced]
      required:
        - id
        - nominationId
        - raceId
        - runnerId
        - trackName
        - raceDate
        - horseName
        - finishPosition
        - fieldSize
        - winStake
        - placeStake
        - totalOutlay
        - netResult
        - outcome
...