# lib/api-zod/src/generated/types/filterResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aussie Horse Win API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FilterResult {
  rule: string;
  passed: boolean;
  message: string;
}