# lib/api-zod/src/generated/types/resultInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aussie Horse Win API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResultInput {
  runnerId: number;
  finishPosition: number;
  /** @nullable */
  actualWinReturn?: number | null;
  /** @nullable */
  actualPlaceReturn?: number | null;
}