# lib/api-zod/src/generated/types/runner.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aussie Horse Win API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FilterResult } from './filterResult';
import type { RunnerSpeedMapPosition } from './runnerSpeedMapPosition';

export interface Runner {
  id: number;
  raceId: number;
  horseName: string;
  barrierNumber: number;
  speedMapPosition: RunnerSpeedMapPosition;
  winOdds: number;
  placeOdds: number;
  /** @nullable */
  jockey?: string | null;
  /** @nullable */
  trainer?: string | null;
  passed: boolean;
  filterResults: FilterResult[];
}